requirements.txt
gradio==3.5
google-auth==2.21.0
google-auth-oauthlib==0.7.0
google-auth-httplib2==0.1.0
google-api-python-client==2.92.0
openai>=0.27.0
pandas==2.1.0
